refactor: refactor academics page and optimized the api call
